pyproject.toml
[project]
name = "palettegen"
version = "0.1.0"
description = "Palette generator with K-Nearest Clustering."
dependencies = ["flask", "numpy", "Pillow", "scikit-learn"]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["palettegen"]
